# 3DFileViewer+LibGL+Lib*GPU: Small performance improvements