--- a/Userland/Libraries/LibGPU/Device.h
+++ b/Userland/Libraries/LibGPU/Device.h
@@ -39,7 +39,7 @@ public:
 
     virtual DeviceInfo info() const = 0;
 
-    virtual void draw_primitives(PrimitiveType, FloatMatrix4x4 const& model_view_transform, FloatMatrix4x4 const& projection_transform, Vector<Vertex>& vertices) = 0;
+    virtual void draw_primitives(PrimitiveType, Vector<Vertex>& vertices) = 0;
     virtual void resize(Gfx::IntSize min_size) = 0;
     virtual void clear_color(FloatVector4 const&) = 0;
     virtual void clear_depth(DepthType) = 0;
@@ -59,6 +59,8 @@ public:
     virtual NonnullRefPtr<Image> create_image(PixelFormat const&, u32 width, u32 height, u32 depth, u32 max_levels) = 0;
     virtual ErrorOr<NonnullRefPtr<Shader>> create_shader(IR::Shader const&) = 0;
 
+    virtual void set_model_view_transform(FloatMatrix4x4 const&) = 0;
+    virtual void set_projection_transform(FloatMatrix4x4 const&) = 0;
     virtual void set_sampler_config(unsigned, SamplerConfig const&) = 0;
     virtual void set_light_state(unsigned, Light const&) = 0;
     virtual void set_material_state(Face, Material const&) = 0;
@@ -68,7 +70,7 @@ public:
 
     virtual RasterPosition raster_position() const = 0;
     virtual void set_raster_position(RasterPosition const& raster_position) = 0;
-    virtual void set_raster_position(FloatVector4 const& position, FloatMatrix4x4 const& model_view_transform, FloatMatrix4x4 const& projection_transform) = 0;
+    virtual void set_raster_position(FloatVector4 const& position) = 0;
 
     virtual void bind_fragment_shader(RefPtr<Shader>) = 0;
 };