* 3DFileViewer: Show the correct frame rate, try to run at 60 FPS
* LibGL+Lib*GPU: Optimize normal transformation calculation
* LibGL & LibSoftGPU: Optimize vertex appends by performing safe unchecked appends

On my machine this yields some small FPS improvements:

| Game | Old FPS | New FPS |
| --- | --- | ---
| glquake | 20.4 | 21.4 (+5%) |
| quake3 | 10.7 | 11.1 (+4%) |
| tuxracer | 25.6 | 26.1 (+2%) |